refactor(15-01): delete timestamp_aligner.py, fix imports
- Delete cesar/timestamp_aligner.py (WhisperX handles alignment)
- Move TranscriptionSegment dataclass to transcriber.py
- Move format_timestamp function to transcript_formatter.py
- Update formatter to use duck typing (List[Any]) for WhisperXSegment compat
- Verify formatter works with WhisperXSegment directly
